+++ src/scriptlets/scriptlets-list.js
@@ -55,3 +55,4 @@ export * from './trusted-set-cookie-reload';
 export * from './trusted-replace-fetch-response';
 export * from './trusted-set-local-storage-item';
 export * from './trusted-set-constant';
+export * from './inject-css-in-shadow-dom';