+++ tests/scriptlets/index.test.js
@@ -50,3 +50,4 @@ import './trusted-set-cookie.test';
 import './trusted-replace-fetch-response.test';
 import './trusted-set-local-storage-item.test';
 import './trusted-set-constant.test';
+import './inject-css-in-shadow-dom.test';